:recycle: Refactored whole application for comments, useless code, best practices
