+++ music-app/src/assets/music-logo-new.png
