+++ music-app/src/assets/player-stop.png
