+++ music-app/src/assets/skip-right.png
