+++ music-app/src/components/Footer.js
@@ -1,8 +1,10 @@
 import React from 'react';
+
 import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
+
 import '../App.css';
 // import { textAlign } from '@material-ui/system';
