+++ music-app/src/components/MusicPlayer.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 
-import { Grid, Typography } from '@material-ui/core';
+import { Grid } from '@material-ui/core';
 
-import SkipLeft from '../images/skip-left.png';
-import SkipRight from '../images/skip-right.png';
-import Pause from '../images/player-stop.png';
-import Play from '../images/player-start.png';
-import loadingSpinner from '../images/lava-lamp-preloader.svg';
+import SkipLeft from '../assets/skip-left.png';
+import SkipRight from '../assets/skip-right.png';
+import Pause from '../assets/player-stop.png';
+import Play from '../assets/player-start.png';
+import loadingSpinner from '../assets/lava-lamp-preloader.svg';
 
 class MusicPlayer extends Component {
   constructor(props) {
@@ -72,38 +72,31 @@ class MusicPlayer extends Component {
   }
 
   createEventHandlers() {
-    // problem setting up the player
     this.player.on('initialization_error', e => {
       console.error(e);
     });
-    // problem authenticating the user.
-    // either the token was invalid in the first place,
-    // or it expired (it lasts one hour)
+
     this.player.on('authentication_error', e => {
       console.error(e);
       this.setState({ loggedIn: false });
     });
-    // currently only premium accounts can use the API
+
     this.player.on('account_error', e => {
       console.error(e);
     });
-    // loading/playing the track failed for some reason
+
     this.player.on('playback_error', e => {
       console.error(e);
     });
 
-    // Playback status updates
     this.player.on('player_state_changed', state => {
       this.onStateChanged(state);
       this.getCurrentSong();
     });
 
-    // Ready
     this.player.on('ready', async data => {
       let { device_id } = data;
-      console.log('Let the music play on!');
-      // set the deviceId variable, then let's try
-      // to swap music playback to *our* player!
+
       await this.setState({ deviceId: device_id, loggedIn: true });
       this.transferPlaybackHere();
     });
@@ -112,21 +105,18 @@ class MusicPlayer extends Component {
   checkForPlayer() {
     const { token } = this.state;
 
-    // if the Spotify SDK has loaded
     if (window.Spotify !== undefined) {
-      // cancel the interval
       clearInterval(this.playerCheckInterval);
-      // create a new player
+
       this.player = new window.Spotify.Player({
         name: 'Music Meteorologist Spotify Player',
         getOAuthToken: cb => {
           cb(token);
         },
       });
-      // set up the player's event handlers
+
       this.createEventHandlers();
 
-      // finally, connect!
       this.player.connect();
     }
   }
@@ -147,8 +137,8 @@ class MusicPlayer extends Component {
         console.log(err);
       });
   }
-  // SDK Player Song playback controls
 
+  // SDK Player Song playback controls
   onPrevClick() {
     this.player.previousTrack();
   }
@@ -162,7 +152,7 @@ class MusicPlayer extends Component {
   }
 
   transferPlaybackHere() {
-    const { deviceId, token } = this.state;
+    const { token } = this.state;
     fetch(
       `https://api.spotify.com/v1/me/player/play?device_id=${this.state.deviceId}`,
       {
@@ -187,15 +177,7 @@ class MusicPlayer extends Component {
   }
 
   render() {
-    const {
-      token,
-      loggedIn,
-      trackName,
-      artistName,
-      albumName,
-      error,
-      playing,
-    } = this.state;
+    const { trackName, artistName, albumName, error, playing } = this.state;
 
     return (
       <Grid container direction='column' justify='center' alignItems='center'>
@@ -207,7 +189,7 @@ class MusicPlayer extends Component {
           )}
           {window.Spotify !== undefined && this.state.imageUrl !== '' && (
             <div className='album-art'>
-              <h4>Now Playing</h4>
+              <h4 style={{ textAlign: 'center' }}>Now Playing</h4>
               <img src={this.state.imageUrl} alt='album-art' />
             </div>
           )}
@@ -216,10 +198,11 @@ class MusicPlayer extends Component {
         {error && <p>Error: {error}</p>}
 
         <Grid item>
-          <p>Artist: {artistName}</p>
+          <div>Artist: {artistName}</div>
           <p>Track: {trackName}</p>
           <p>Album: {albumName}</p>
         </Grid>
+
         <Grid
           container
           direction='row'
