+++ music-app/src/components/NavBar.js
@@ -1,15 +1,14 @@
 import React from 'react';
-import Button from '@material-ui/core/Button';
-import Auth from './Auth';
-// import { flexbox } from '@material-ui/system';
+
 import { Grid, Typography } from '@material-ui/core';
-import MusicLogo from '../images/music-logo-new.png';
+import Button from '@material-ui/core/Button';
+
+import MusicLogo from '../assets/music-logo-new.png';
 
 class NavBar extends React.Component {
   homeButton = e => {
     e.preventDefault();
-    window.location.href = '/'; // need to figure out how to pass down props here
-    // this.history.push("/")
+    window.location.href = '/';
   };
   render() {
     return (
@@ -31,10 +30,6 @@ class NavBar extends React.Component {
         <Grid item>
           <img src={MusicLogo} style={{ height: 70 }} />
         </Grid>
-
-        {/* <Grid item>
-          <Auth />
-        </Grid> */}
       </Grid>
     );
   }
