+++ music-app/src/index.css
@@ -5,7 +5,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url('background.jpg');
+  background-image: url('/assets/background.jpg');
   background-size: 100%;
   color: white;
 }
